requirements.txt
streamlit
yt-dlp
pydub
requests
groq
google-generativeai

llama-index-core==0.10.58
llama-index-readers-file==0.1.30
llama-index-llms-nvidia==0.1.4
llama-index-embeddings-nvidia==0.1.4
llama-index-vector-stores-milvus==0.1.20
pymilvus==2.4.4
gradio==4.37.2